Update GPU benchmarks to use bfloat16 data type
- Change from float16 to bfloat16 (widely used in modern deep learning)
- bfloat16 offers same range as float32 with float16 bandwidth/memory
- Update CUDA extension to support bfloat16: use AT_DISPATCH_FLOATING_TYPES_AND2
  to include both torch::kHalf and torch::kBFloat16
- Update plot title and peak performance labels from FP16 to BF16
- All GPU implementations now validated with bfloat16

Performance at 4096^3 with BF16:
- torch_matmul (cuBLAS): 143,635 GFLOPS
- cuda_tiled_32x32: 5,997 GFLOPS
- cuda_naive_ijk_jx: 5,430 GFLOPS
- cuda_naive_ijk: 1,280 GFLOPS

bfloat16 is preferred over float16 for training because:
- Same exponent range as float32 (no dynamic range loss)
- Better numerical stability during gradient updates
- Widely supported by modern GPUs (Ampere+)

+++ benchmarks/bench_gpu.py
@@ -90,8 +90,8 @@ def run(impls, sizes):
                 continue
 
             # Create tensors directly on GPU
-            A_gpu = torch.randn(M, K, dtype=torch.float16, device='cuda')
-            B_gpu = torch.randn(K, N, dtype=torch.float16, device='cuda')
+            A_gpu = torch.randn(M, K, dtype=torch.bfloat16, device='cuda')
+            B_gpu = torch.randn(K, N, dtype=torch.bfloat16, device='cuda')
 
             # Validate result before benchmarking
             try:
